feat: add .nvmrc file to specify Node.js version
This commit adds a new file called .nvmrc that specifies the version of
Node.js required for this project. This is useful for developers who use
Node Version Manager (nvm) to manage their Node.js installations and want
to ensure that the correct version is used for this project. The version
specified in this file is 16.